+++ src/components/SearchSettings/SettingsSwitcher.tsx
@@ -42,10 +42,12 @@ export function SettingsSwitcherItem({ name, defaultValue, icon, value, onClick 
 
     return (
         <>
-            <input type="radio" id={`theme-switcher-${value}`} name={name} value={value} defaultChecked={defaultValue === value} />
-            <label onClick={onClick} htmlFor={`theme-switcher-${value}`} aria-label={value} >
-                {icon}
-            </label>
+            <div className="flex">
+                <input type="radio" id={`theme-switcher-${value}`} name={name} value={value} defaultChecked={defaultValue === value} />
+                <label onClick={onClick} htmlFor={`theme-switcher-${value}`} aria-label={value} >
+                    {icon}
+                </label>
+            </div>
         </>
     )
 }